+++ database.js
@@ -1,5 +1,4 @@
 import mongoose from "mongoose";
-import sequence from "mongoose-sequence";
 
 mongoose.connect('mongodb://localhost/disclone', {
   useNewUrlParser: true,
@@ -8,7 +7,6 @@ mongoose.connect('mongodb://localhost/disclone', {
   useFindAndModify: false,
 });
 const db = mongoose.connection;
-const seq = sequence(db);
 
 db.on('error', console.error.bind(console, 'connection error:'));
 db.once('open', function() {
@@ -16,4 +14,3 @@ db.once('open', function() {
 });
 
 export default db;
-export {seq};