+++ models/channel.schema.js
@@ -1,0 +1,8 @@
+import mongoose from "mongoose";
+
+export const channelSchema = new mongoose.Schema({
+    name: String,
+    history: {type: Array, default: []}
+});
+
+export default mongoose.model('Channel', channelSchema);