+++ routes/user.js
@@ -1,43 +1,21 @@
 import express from "express";
 import User from '../models/user.schema.js';
-import Request from '../models/request.schema.js';
 import bcrypt from 'bcrypt';
-import {deleteRequest} from './request.js';
 
 const router = express.Router();
 const saltRounds = 10;
 
-function getTimeWithLeadingZeros() {
-    const d = new Date();
-    const hours = (d.getHours() > 10 ? '' : '0') + d.getHours();
-    const minutes = (d.getMinutes() > 10 ? '' : '0') + d.getMinutes();
-    const seconds = (d.getSeconds() > 10 ? '' : '0') + d.getSeconds();
-    return `${hours}:${minutes}:${seconds}`;
-}
-
-
-
-router.use(function (req, res, next) {
-    const date = getTimeWithLeadingZeros();
-    const host = req.headers.host;
-    const method = req.method;
-    const url = req.originalUrl;
-    const request = `${host} ${method} ${url}`;
-    console.log(`[${date}]: ${request}`);
-    next();
-});
-
 router.get('/u/:uid', function(req, res) {
     if (!req.params.uid) return res.status(400).send("No id");
-    const uid = parseInt(req.params.uid);
+    const uid = req.params.uid;
 
-    User.findOne({uid}, function(err, user) {
+    User.findOne({_id: uid}, function(err, user) {
         if (err) {
             console.error(err);
             return res.status(400).send(err);
         }
 
-        return res.status(200).json({uid: user.uid, username: user.username});
+        return res.status(200).json({_id: user._id, username: user.username});
     });
 })
 
@@ -45,7 +23,7 @@ router.post('/register', function(req, res) {
     if (!req.body.username || !req.body.password) return res.status(400).send("No data");
     const username = req.body.username.toLowerCase();
     const password = bcrypt.hashSync(req.body.password, saltRounds);
-    const user = new User({username, password});
+    const user = new User({username, password, friends: []});
 
     user.save(function (err, user) {
         if (err) {
@@ -54,7 +32,7 @@ router.post('/register', function(req, res) {
         }
 
         console.log(`Registererd user ${username}`);
-        return res.status(200).json({uid: user.uid, username: user.username, friends: user.friends});
+        return res.status(200).json({_id: user._id, username: user.username, friends: user.friends});
     })
 });
 
@@ -62,7 +40,7 @@ router.post('/login', function(req, res) {
     if (!req.body.username || !req.body.password) return res.status(400).send("No data");
     const {username, password} = req.body;
 
-    User.findOne({username: username.toLowerCase()}, '-_id -__v', function(err, user) {
+    User.findOne({username: username.toLowerCase()}, '-__v', function(err, user) {
         if (err) return res.status(400).send(err);
         if (!user) return res.status(400).send("No user");
         if (!bcrypt.compareSync(password, user.password)) return res.status(400).send("Invalid password");
@@ -77,7 +55,7 @@ router.post('/members', function(req, res) {
     const {members, server} = req.body;
     if (!members || !server) return res.status(400).send("No data");
     
-    User.find({uid: members}).select("username uid -_id").exec(function(err, users) {
+    User.find({_id: members}).select("username").exec(function(err, users) {
         if (err) {
             console.error(err);
             return res.status(400).send(err);
@@ -92,15 +70,15 @@ router.post('/friends', function(req, res) {
     const {friends} = req.body;
     if (!friends) return res.status(400).send("No data");
 
-    User.find({uid: friends}, 'username uid -_id', function(err, users) {
+    User.find({_id: friends}, 'username', function(err, users) {
         if (err) return res.status(400).send(err);
-
+        console.log(users);
         return res.status(200).json(users);
     });
 });
 
 async function checkIfFriends(uid, fuid) {
-    const uQuery = User.findOne({uid});
+    const uQuery = User.findOne({_id: uid});
     const uRes = await uQuery.exec();
 
     if (!uRes) return {error: "No user"};
@@ -111,24 +89,28 @@ async function checkIfFriends(uid, fuid) {
 export async function addFriends(uid, fuid) {
     if (!uid || !fuid) return { error: "No user ids" };
 
-    uid = parseInt(uid);
-    fuid = parseInt(fuid);
-    if (isNaN(uid) || isNaN(fuid)) return { error: "No valid user ids" };
-
     const friendsCheck = await checkIfFriends(uid, fuid);
     if (friendsCheck.error) return friendsCheck;
 
-    const rQuery = User.findOneAndUpdate({uid}, { $push: { friends: fuid} });
-    const sQuery = User.findOneAndUpdate({uid: fuid}, { $push: { friends: uid} });
+    const rQuery = User.findOneAndUpdate({_id: uid}, { $push: { friends: fuid} });
+    const sQuery = User.findOneAndUpdate({_id: fuid}, { $push: { friends: uid} });
     const rRes = await rQuery.exec();
     const sRes = await sQuery.exec();
 
     if (!rRes.friends.includes(fuid)) rRes.friends.push(fuid);
     if (!sRes.friends.includes(uid)) sRes.friends.push(uid);
     
-    return { error: false, receiver: {uid: sRes.uid, username: sRes.username}, sender: {uid: rRes.uid, username: rRes.username} };
+    return { error: false, receiver: {_id: sRes._id, username: sRes.username}, sender: {_id: rRes._id, username: rRes.username} };
 }
 
+export async function deleteFriend(uid, fuid) {
+    if (!uid || !fuid) return { error: "No user ids" };
+
+    const userQuery = User.updateMany({_id: [uid, fuid]}, {$pullAll: {friends: [uid, fuid]}});
+    const userRes = await userQuery.exec();
+
+    return userRes;
+}
 
 
 export default router;